refac(virtual): virtual content DRYing
